requirements/requirements.txt
appdirs==1.4.4
arrow==1.1.1
asgiref==3.4.1
binaryornot==0.4.4
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cookiecutter==1.7.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.10.0
defusedxml==0.7.1
Django==2.2.20
django-allauth==0.45.0
django-analytical==3.0.0
django-cors-headers==3.8.0
django-crispy-forms==1.12.0
django-debug-toolbar==3.2.2
django-extra-views==0.14.0
django-hashid-field==3.3.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.2
et-xmlfile==1.1.0
ez-setup==0.9
form==0.0.1
future==0.18.2
hashids==1.3.1
idna==3.2
itypes==1.2.0
jdcal==1.4.1
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
oauthlib==3.1.1
olefile==0.46
openapi-codec==1.3.2
openpyxl==3.0.7
packaging==21.0
Pillow==8.3.1
poyo==0.5.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==2.4.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
text-unidecode==1.3
uritemplate==3.0.1
urllib3==1.26.6
whichcraft==0.6.1
